refactor: 입력된 자동차 이름에 대한 검증을 CarNameDto -> InputView 로 수정
