+++ src/main/java/view/InputView.java
@@ -8,13 +8,18 @@ import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
 
+import static domain.Car.BLANK;
+import static domain.Car.MAX_NAME_LENGTH;
+
 public class InputView {
 
     public List<CarNameDto> sendCarsName() {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
 
         Scanner scanner = new Scanner(System.in);
-        return Arrays.stream(scanner.next().split(","))
+
+        return Arrays.stream(scanner.nextLine().split(","))
+                .peek(this::validate)
                 .map(CarNameDto::new)
                 .collect(Collectors.toList());
     }
@@ -24,5 +29,24 @@ public class InputView {
 
         Scanner scanner = new Scanner(System.in);
         return new TryCount(scanner.nextInt());
+    }
+
+    private void validate(String name) {
+        validateToContainBlank(name);
+        validateLength(name);
+    }
+
+    private void validateLength(String name) {
+        if (name.length() > MAX_NAME_LENGTH) {
+            throw new IllegalArgumentException("이름은 " + MAX_NAME_LENGTH + "자 이하여야 합니다. " +
+                    "입력값 : "  + name);
+        }
+    }
+
+    private void validateToContainBlank(String name) {
+        if (name.contains(BLANK)) {
+            throw new IllegalArgumentException("이름에 공백을 포함할 수 없습니다." +
+                    "입력값 : " + name);
+        }
     }
 }
